refactor: move PlanarTileProcessing internals to PlanarLayer
To continue the removal of some Processing files in Process, all
contents of PlanarTileProcessing have been moved to PlanarLayer.
